WIP: Extract base node behaviour into a separate kg-default-nodes package
refs https://github.com/TryGhost/Team/issues/2225

Node definitions are currently duplicated in the editor and in the renderer. We want to add URL transformation details to nodes which could mean a further duplication unless we find a way to have core node logic (payload getters/setters, import, render, etc) separated in a Node friendly way for use in other server-side packages.

- added `kg-default-nodes` package as a central location for all base node logic, the Lexical equivalent to our Mobiledoc `kg-default-cards` package`
  - extracted all generic image card logic into the new package
  - split out the DOM import function to a separate file to make unit testing easier (this is due to be expanded upon later so keeping it separate will help prevent the node file growing too big and keeping it focused on payload data)
- added sourcemap generation to the demo build config to make debugging easier
