+++ packages/kg-default-nodes/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+    plugins: ['ghost'],
+    extends: [
+        'plugin:ghost/node'
+    ],
+    parseOptions: {
+        sourceType: 'module'
+    }
+};
