+++ packages/kg-default-nodes/README.md
@@ -1,0 +1,35 @@
+# Koenig Default Nodes
+
+Lexical node definitions for the default nodes used in Ghost&#39;s Koenig editor
+
+## Install
+
+`npm install @tryghost/kg-default-nodes --save`
+
+or
+
+`yarn add @tryghost/kg-default-nodes`
+
+## Usage
+
+
+## Develop
+
+This is a monorepo package.
+
+Follow the instructions for the top-level repo.
+1. `git clone` this repo & `cd` into it as usual
+2. Run `yarn` to install top-level dependencies.
+
+
+
+## Test
+
+- `yarn lint` run just eslint
+- `yarn test` run lint and tests
+
+
+
+# Copyright & License 
+
+Copyright (c) 2013-2022 Ghost Foundation - Released under the [MIT license](LICENSE).
