+++ packages/kg-default-nodes/index.js
@@ -1,0 +1,1 @@
+export * from './lib/kg-default-nodes';
