+++ packages/kg-default-nodes/lib/kg-default-nodes.js
@@ -1,0 +1,1 @@
+export * from './image/ImageNode';
