+++ packages/koenig-lexical/package.json
@@ -39,6 +39,7 @@
     "@lexical/react": "^0.6.0",
     "@lexical/selection": "^0.6.0",
     "@lexical/utils": "^0.6.0",
+    "@tryghost/kg-default-nodes": "^0.0.0",
     "lexical": "^0.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
